requirements.txt
google-genai>=1.30.0
pydantic>=2.8.0
streamlit>=1.37.0
PyMuPDF>=1.24.0
python-dotenv>=1.0.1
fastapi>=0.115.0
uvicorn[standard]>=0.30.0